Move pending transaction logic in the parent fragment
